体重を読み込んで表示したい
