+++ pppppp/LoginViewController.swift
@@ -60,7 +60,6 @@ class LoginViewController: UIViewController, UITextFieldDelegate {
                 alert.addAction(UIAlertAction(title: "OK", style: .default, handler: nil))
                 self.present(alert, animated: true, completion: nil)
             }
-            //            }
         }
     }
