+++ pppppp/WeightViewController.swift
@@ -27,13 +27,12 @@ class WeightViewController: UIViewController, UITextFieldDelegate {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        
         let types = Set([typeOfBodyMass])
         let healthStore = HKHealthStore()
         healthStore.requestAuthorization(toShare: types, read: types, completion: { success, error in
             print(success)
-//            print(error)
         })
+        
         self.weightTextField?.delegate = self
         read()
     }
@@ -49,7 +48,7 @@ class WeightViewController: UIViewController, UITextFieldDelegate {
             if success {
                 NSLog("HealthKit保存成功!")
             } else {
-                print("HealthKit保存成功!")
+                print("HealthKitできませんでした。")
             }
         })
         
@@ -73,15 +72,61 @@ class WeightViewController: UIViewController, UITextFieldDelegate {
         }
     }
     
-    //データを取得
+        //データを取得
+//    func read() {
+//        let query = HKSampleQuery(sampleType: typeOfBodyMass, predicate: nil, limit: 1, sortDescriptors: nil) { (query, results, error) in
+//            if results is [HKQuantitySample] {
+//                print("\(query)")
+//                print("データを取得しました")
+//
+//            }
+//        }
+//        myHealthStore.execute(query)
+//    }
+    
     func read() {
-        let query = HKSampleQuery(sampleType: typeOfBodyMass, predicate: nil, limit: 1, sortDescriptors: nil) { (query, results, error) in
+        
+        let query = HKSampleQuery(sampleType: typeOfBodyMass, predicate: nil, limit: Int(1.0), sortDescriptors: nil) { (query, results, error) in
             if results is [HKQuantitySample] {
-                print("データを取得しました")
+                if results is [HKQuantitySample] {
+                    // 取得したデータを格納
+                    print("体重ううううううは\(String(describing: results))")
+                }
             }
         }
         myHealthStore.execute(query)
     }
+    
+//    func read(){
+//            // 取得する期間を設定
+//            let dateformatter = DateFormatter()
+//            dateformatter.dateFormat = "yyyy/MM/dd"
+//            let startDate = dateformatter.date(from: "2000/1/1")
+//            let endDate = dateformatter.date(from: "2022/06/21")
+//
+//            // 取得するデータを設定
+//            let typeOfWeight = HKObjectType.quantityType(forIdentifier: HKQuantityTypeIdentifier.bodyMass)
+//            let statsOptions: HKStatisticsOptions = [HKStatisticsOptions.discreteMin, HKStatisticsOptions.discreteMax]
+//
+//            let predicate = HKQuery.predicateForSamples(withStart: startDate, end: endDate, options: HKQueryOptions.strictStartDate)
+//            let query = HKStatisticsQuery(quantityType: typeOfWeight!, quantitySamplePredicate: predicate, options: statsOptions, completionHandler: { (query, result, error) in
+//                if let e = error {
+//                    print("Error: \(e.localizedDescription)")
+//                    return
+//                }
+//                DispatchQueue.main.async {
+//                    guard let r = result else {
+//                        return
+//                    }
+//                    let min = r.minimumQuantity()
+//                    let max = r.maximumQuantity()
+//                    if min != nil && max != nil {
+//                        print("\(r.startDate) : \(r.endDate) 最小:\(min!) 最大:\(max!)")
+//                    }
+//                }
+//            })
+//            myHealthStore.execute(query)
+//        }
     
     func textFieldShouldReturn(_ textField: UITextField) -> Bool {
         weightTextField.resignFirstResponder()
